// Template Driven Forms/Porf2/curso-formularios-com-angular-reactive-e-template-driven-forms/src/app/features/reactive-forms/13.json-form-example/json-form/interfaces/form.schema.interface.ts
import { ValidatorFn } from '@angular/forms';
import { FieldType } from '../enums/field-type.enum';

export interface JsonFormSchema {
  submitLabel: string;
  fields: JsonFormFieldSchema[];
  validators?: ValidatorFn[];
}

export interface JsonFormFieldSchema {
  name: string;
  label: string;
  placeholder?: string;
  intialValue: unknown;
  required: boolean;
  type: FieldType;
  options?: { value: unknown; label: string }[];
  validators?: ValidatorFn[];
}